+++ arandomness/argparse/__init__.py
@@ -22,6 +22,7 @@ Copyright:
 
 from .CheckThreads import CheckThreads
 from .CopyRight import CopyRight
+from .Open import Open
 from .ParseCommas import ParseCommas
 
 __author__ = 'Alex Hyer'
@@ -29,4 +30,4 @@ __email__ = 'theonehyer@gmail.com'
 __license__ = 'GPLv3'
 __maintainer__ = 'Alex Hyer'
 __status__ = 'Production/Stable'
-__version__ = '1.0.4'
+__version__ = '1.0.5'